Prevent MSVC compiler warning in MemoryStreamWriter::Seek
 - Add static_cast when setting this->offset to remove warning.

+++ StreamWriter.cpp
@@ -82,7 +82,7 @@ void MemoryStreamWriter::Seek(uint64_t offset)
 		throw std::runtime_error("Change in offset places read position outside bounds of buffer.");
 	}
 
-	this->offset = offset;
+	this->offset = static_cast<size_t>(offset);
 }
 
 void MemoryStreamWriter::SeekRelative(int64_t offset)
